FOV w/ cubic regression, anti-macro option

+++ src/mudclient.h
@@ -222,9 +222,9 @@ typedef struct mudclient mudclient;
 #include "ui/welcome.h"
 #include "ui/wilderness-warning.h"
 
-extern int test_x;
-extern int test_y;
-extern int test_z;
+extern float test_x;
+extern float test_y;
+extern float test_z;
 extern int test_yaw;
 extern int test_colour;
 extern int test_fade;
